requirements.txt
schedule
python-telegram-bot
nsepy
pandas
numpy
gspread_pandas
nsetools